Packaged code for PyPI by sashanje
