+++ sashanje_diffusion2d.egg-info/PKG-INFO
@@ -1,3 +1,16 @@
+Metadata-Version: 2.1
+Name: sashanje_diffusion2d
+Version: 0.0.3
+Summary: Solves the diffusion equation in 2D over a square domain
+Author: Sashan Samarajeewa
+Project-URL: Homepage, https://github.com/Simulation-Software-Engineering/diffusion2D
+Keywords: diffusion,simulation
+Classifier: Programming Language :: Python :: 3
+Description-Content-Type: text/markdown
+License-File: LICENSE
+Requires-Dist: numpy
+Requires-Dist: matplotlib
+
 # diffusion2D
 
 ## Instructions for students
