+++ sashanje_diffusion2d/output.py
@@ -1,0 +1,12 @@
+def create_plot(n, plt, u, ax, T_cold, T_hot, dt):
+        im = ax.imshow(u.copy(), cmap=plt.get_cmap('hot'), vmin=T_cold, vmax=T_hot)  # image for color bar axes
+        ax.set_axis_off()
+        ax.set_title('{:.1f} ms'.format(n * dt * 1000))
+        return im
+
+def output_plots(fig, plt, im):
+    fig.subplots_adjust(right=0.85)
+    cbar_ax = fig.add_axes([0.9, 0.15, 0.03, 0.7])
+    cbar_ax.set_xlabel('$T$ / K', labelpad=20)
+    fig.colorbar(im, cax=cbar_ax)
+    plt.show()
